@@ -1,7 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
 import thunk from 'redux-thunk';
 import { homeReducer } from './home/home';
-import { detailsReducer } from './details/details';
+import detailsReducer from './details/details';
 
 const store = configureStore({
   reducer: {
