@@ -1,7 +1,8 @@
 import axios from 'axios';
 
 const apiKey = '?apikey=cec71bb490150a3c8d68e3c66b2f0a02';
-const stockApi = `https://financialmodelingprep.com/api/v3/stock/list${apiKey}`;
+//                https://financialmodelingprep.com/api/v3/financial-statement-symbol-lists?apikey=cec71bb490150a3c8d68e3c66b2f0a02
+const stockApi = `https://financialmodelingprep.com/api/v3/financial-statement-symbol-lists${apiKey}`;
 
 const GET_STOCKS = 'GET_STOCKS';
 const GET_STOCKS_SUCCESS = 'GET_STOCKS_SUCCESS';
@@ -11,10 +12,13 @@ const getStocks = () => async (dispatch) => {
   dispatch({ type: GET_STOCKS });
   try {
     const response = await axios.get(stockApi);
-    dispatch({ type: GET_STOCKS_SUCCESS, payload: response.data });
+    const { data } = response;
+    dispatch({ type: GET_STOCKS_SUCCESS, payload: data });
   } catch (error) {
     dispatch({ type: GET_STOCKS_ERROR, payload: error.message });
   }
 };
 
-export { getStocks, GET_STOCKS, GET_STOCKS_ERROR, GET_STOCKS_SUCCESS };
+export {
+  getStocks, GET_STOCKS, GET_STOCKS_ERROR, GET_STOCKS_SUCCESS,
+};
